feat(builder): add missing id() on entity builders
Also create the intermediary AbstractBuilder class between Builder and
ExtensibleBuilder. This way:
- Builder, AbstractBuilder and ExtensibleBuilder are completely
generics, and so can be reused without any business context.
- An EntityBuilder is now an abstract ExtensibleBuilder.
