+++ src/main/java/de/fraunhofer/iosb/ilt/sta/model/builder/api/AbstractObservedPropertyBuilder.java
@@ -8,12 +8,12 @@ import java.net.URI;
 import java.util.List;
 
 /**
- * Base class for any {@link AbstractEntityBuilder} of {@link ObservedProperty}
+ * Base class for any {@link EntityBuilder} of {@link ObservedProperty}
  *
  * @param <U> the type of the concrete class that extends this {@link AbstractObservedPropertyBuilder}
  * @author Aurelien Bourdon
  */
-public abstract class AbstractObservedPropertyBuilder<U extends AbstractObservedPropertyBuilder<U>> extends AbstractEntityBuilder<ObservedProperty, U> {
+public abstract class AbstractObservedPropertyBuilder<U extends AbstractObservedPropertyBuilder<U>> extends EntityBuilder<ObservedProperty, U> {
 
     @Override
     protected ObservedProperty newBuildingInstance() {
