+++ src/main/java/de/fraunhofer/iosb/ilt/sta/model/builder/api/BuildingException.java
@@ -7,12 +7,16 @@ package de.fraunhofer.iosb.ilt.sta.model.builder.api;
  */
 public class BuildingException extends RuntimeException {
 
+    public BuildingException(final String message) {
+        super(message);
+    }
+
     public BuildingException(final Throwable cause) {
         super(cause);
     }
 
-    public BuildingException(final String message) {
-        super(message);
+    public BuildingException(final String message, final Throwable cause) {
+        super(message, cause);
     }
 
 }
